requirements.txt
biopython
pandas
tensorflow
scikit-learn
spacy 